Add db store backed table entry for ou user table (#93)
This ou user table will also allow mapping the relevant role for a user
in OU, currently we will be starting with static system defined roles
of admin, default and auditor, eventually custom role definitions based
on policy will be provided out of the system based on the resources
available

Also handle offsets and limits as part of this change

+++ pkg/server/tenant-user.go
@@ -67,7 +67,7 @@ func (s *TenantUserApiServer) GetUsers(ctx context.Context, req *api.TenantUsers
 		Items: []*api.TenantUserListEntry{},
 	}
 
-	users, err := s.userTbl.GetByTenant(ctx, req.Tenant, int64(req.Offset), int64(req.Limit))
+	users, err := s.userTbl.GetByTenant(ctx, req.Tenant, req.Offset, req.Limit)
 	if err != nil && !errors.IsNotFound(err) {
 		log.Printf("failed to fetch users for tenant %s: %s", req.Tenant, err)
 		return nil, status.Errorf(codes.Internal, "Something went wrong, Please try again later")
